pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tenant_legal_guidance"
version = "0.1.0"
description = "A system for scraping and managing tenant legal information using knowledge graphs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.5",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.2.1",
    "spacy>=3.7.2",
    "networkx>=3.2.1",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "python-arango>=7.5.8",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.3",
    "transformers>=4.37.2",
    "torch>=2.2.0",
    "torch-geometric>=2.4.0",
    "pytesseract>=0.3.10",
    "pdf2image>=1.16.3",
    "python-docx>=1.0.1",
    "aiohttp>=3.9.1",
    "aiofiles>=0.7.0",
    "PyPDF2>=3.0.0",
    "markdown>=3.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=24.1.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "ruff>=0.2.1",
    "en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=tenant_legal_guidance"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.hatch.metadata]
allow-direct-references = true 

[dependency-groups]
dev = [
    "httpx>=0.28.1",
]
